engine: Move TriggerInputs to dsl
