+++ tracecat/dsl/workflow.py
@@ -60,7 +60,7 @@ with workflow.unsafe.imports_passed_through():
         TracecatValidationError,
     )
     from tracecat.types.validation import ValidationResult
-    from tracecat.webhooks.models import TriggerInputs
+    from tracecat.dsl.models import TriggerInputs
     from tracecat.workflow.management.definitions import (
         get_workflow_definition_activity,
     )
